fix: no warning modal is displayed when unapproving the exported expense report
